Basic Faction class created
The OOP class for Faction has been created.
It has a proper constructor, methods to join and
leave the faction and a few other properties that
aren't used yet but will be soon.
It's only a little jank and nothing is exposed to
the user yet (I have added no commands).